datastream/requirements.txt
python-dotenv==0.17.0
alpaca_trade_api==1.1.0
psycopg2-binary==2.8.6